backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==4.0.2
django-cors-headers==3.11.0
django-solo==2.0.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
openpyxl==3.1.2
oracledb==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
requests==2.27.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.1.1
urllib3==1.26.8
zope.interface==5.4.0